Pull packages from nuget. Layout changes.
Fix CRM2015 exception

In CRM2015 there is no regardingobjectid. Use ConnectionDetail to check this is modify the initial fetch query.

Show audit details in split panel instead of popup
