+++ BDK.XrmToolBox/BDK.XrmToolBox.RecycleBin/packages.config
@@ -1,0 +1,14 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<packages>
+  <package id="Microsoft.CrmSdk.CoreAssemblies" version="8.1.0.2" targetFramework="net452" />
+  <package id="Microsoft.CrmSdk.Deployment" version="8.1.0.2" targetFramework="net452" />
+  <package id="Microsoft.CrmSdk.Workflow" version="8.1.0.2" targetFramework="net452" />
+  <package id="Microsoft.CrmSdk.XrmTooling.CoreAssembly" version="8.1.0.2" targetFramework="net452" />
+  <package id="Microsoft.IdentityModel" version="6.1.7600.16394" targetFramework="net452" />
+  <package id="Microsoft.IdentityModel.Clients.ActiveDirectory" version="2.24.304111323" targetFramework="net452" />
+  <package id="Microsoft.Web.Xdt" version="2.1.1" targetFramework="net452" />
+  <package id="MscrmTools.Xrm.Connection" version="1.2016.9.1" targetFramework="net452" />
+  <package id="Newtonsoft.Json" version="9.0.1" targetFramework="net452" />
+  <package id="NuGet.Core" version="2.12.0" targetFramework="net452" />
+  <package id="XrmToolBoxPackage" version="1.2016.11.3" targetFramework="net452" />
+</packages>
